Broke out getThemedLegendGraphicUrl and gave the small legendicon in list the possibility to be transparent.
