+++ apps/client/src/plugins/LayerSwitcher/components/LegendIcon.js
@@ -1,13 +1,19 @@
+import { useTheme } from "@mui/material/styles";
+import { getThemedLegendGraphicUrl } from "../LayerswitcherUtils";
+
 export default function LegendIcon({ url }) {
+  const theme = useTheme();
+  const isDarkMode = theme.palette.mode === "dark";
+
   return (
     <img
       alt="Teckenförklaringsikon"
-      src={url}
+      src={getThemedLegendGraphicUrl(url, isDarkMode)}
       style={{
-        width: "24px",
-        height: "24px",
+        width: "18px",
+        height: "18px",
         marginRight: "5px",
-        marginTop: "4px",
+        marginTop: "8px",
       }}
     ></img>
   );
